+++ src/main/java/cleancode/minesweeper/tobe/GameApplication.java
@@ -1,13 +1,10 @@
 package cleancode.minesweeper.tobe;
 
-import cleancode.minesweeper.tobe.config.GameConfig;
-import cleancode.minesweeper.tobe.gameLevel.Beginner;
-import cleancode.minesweeper.tobe.gameLevel.GameLevel;
-import cleancode.minesweeper.tobe.gameLevel.VeryBeginner;
-import cleancode.minesweeper.tobe.io.ConsoleInputHandler;
-import cleancode.minesweeper.tobe.io.ConsoleOutputHandler;
-import cleancode.minesweeper.tobe.io.InputHandler;
-import cleancode.minesweeper.tobe.io.OutputHandler;
+import cleancode.minesweeper.tobe.minesweeper.config.GameConfig;
+import cleancode.minesweeper.tobe.minesweeper.gameLevel.Beginner;
+import cleancode.minesweeper.tobe.minesweeper.io.ConsoleInputHandler;
+import cleancode.minesweeper.tobe.minesweeper.io.ConsoleOutputHandler;
+import cleancode.minesweeper.tobe.minesweeper.Minesweeper;
 
 public class GameApplication {
     public static void main(String[] args) {
