--- src/main/java/cleancode/minesweeper/tobe/GameException.java
@@ -1,7 +1,0 @@
-package cleancode.minesweeper.tobe;
-
-public class GameException extends RuntimeException {
-    public GameException(String message) {
-        super(message);
-    }
-}
