--- src/main/java/cleancode/minesweeper/tobe/cell/Cell.java
@@ -1,19 +1,0 @@
-package cleancode.minesweeper.tobe.cell;
-
-public interface Cell {
-
-    boolean isLandMine();
-
-    boolean hasLandMineCount();
-
-    CellSnapshot getSnapshot();
-
-    void flag();
-
-    void open();
-
-    boolean isChecked();
-
-    boolean isOpened();
-
-}
