--- src/main/java/cleancode/minesweeper/tobe/cell/CellSnapshotStatus.java
@@ -1,16 +1,0 @@
-package cleancode.minesweeper.tobe.cell;
-
-public enum CellSnapshotStatus {
-
-    EMPTY("빈 셀"),
-    FLAG("깃발"),
-    LANE_MINE("지뢰"),
-    NUMBER("숫자"),
-    UNCHECKED("확인 전");
-
-    private final String description;
-
-    CellSnapshotStatus(String description) {
-        this.description = description;
-    }
-}
