--- src/main/java/cleancode/minesweeper/tobe/config/GameConfig.java
@@ -1,30 +1,0 @@
-package cleancode.minesweeper.tobe.config;
-
-import cleancode.minesweeper.tobe.gameLevel.GameLevel;
-import cleancode.minesweeper.tobe.io.InputHandler;
-import cleancode.minesweeper.tobe.io.OutputHandler;
-
-public class GameConfig {
-
-    private final GameLevel gameLevel;
-    private final InputHandler inputHandler;
-    private final OutputHandler outputHandler;
-
-    public GameConfig(GameLevel gameLevel, InputHandler inputHandler, OutputHandler outputHandler) {
-        this.gameLevel = gameLevel;
-        this.inputHandler = inputHandler;
-        this.outputHandler = outputHandler;
-    }
-
-    public GameLevel getGameLevel() {
-        return gameLevel;
-    }
-
-    public InputHandler getInputHandler() {
-        return inputHandler;
-    }
-
-    public OutputHandler getOutputHandler() {
-        return outputHandler;
-    }
-}
