--- src/main/java/cleancode/minesweeper/tobe/gameLevel/Beginner.java
@@ -1,18 +1,0 @@
-package cleancode.minesweeper.tobe.gameLevel;
-
-public class Beginner implements GameLevel{
-    @Override
-    public int getRowSize() {
-        return 8;
-    }
-
-    @Override
-    public int getColSize() {
-        return 10;
-    }
-
-    @Override
-    public int getLandMineCount() {
-        return 10;
-    }
-}
