--- src/main/java/cleancode/minesweeper/tobe/gameLevel/GameLevel.java
@@ -1,8 +1,0 @@
-package cleancode.minesweeper.tobe.gameLevel;
-
-public interface GameLevel {
-
-    int getRowSize();
-    int getColSize();
-    int getLandMineCount();
-}
