--- src/main/java/cleancode/minesweeper/tobe/io/InputHandler.java
@@ -1,12 +1,0 @@
-package cleancode.minesweeper.tobe.io;
-
-import cleancode.minesweeper.tobe.position.CellPosition;
-import cleancode.minesweeper.tobe.user.UserAction;
-
-public interface InputHandler {
-
-    UserAction getUserActionFromUser();
-
-    CellPosition getCellPositionFromUser();
-
-}
