--- src/main/java/cleancode/minesweeper/tobe/io/OutputHandler.java
@@ -1,24 +1,0 @@
-package cleancode.minesweeper.tobe.io;
-
-import cleancode.minesweeper.tobe.GameBoard;
-import cleancode.minesweeper.tobe.GameException;
-
-public interface OutputHandler {
-
-    void showGameStartComments();
-
-    void showBoard(GameBoard board);
-
-    void showGameWinningComment();
-
-    void showGameLosingComment();
-
-    void showCommentForSelectingCell();
-
-    void showCommentForUserAction();
-
-    void showExceptionMessage(GameException e);
-
-    void showSimpleMessage(String message);
-
-}
