--- src/main/java/cleancode/minesweeper/tobe/io/sign/CellSignProvidable.java
@@ -1,11 +1,0 @@
-package cleancode.minesweeper.tobe.io.sign;
-
-import cleancode.minesweeper.tobe.cell.CellSnapshot;
-
-public interface CellSignProvidable {
-
-    boolean supports(CellSnapshot cellSnapshot);
-
-    String provide(CellSnapshot cellSnapshot);
-
-}
