+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/board/GameStatus.java
@@ -1,0 +1,16 @@
+package cleancode.minesweeper.tobe.minesweeper.board;
+
+public enum GameStatus {
+
+    IN_PROGRESS("진행중"),
+    WIN("승리"),
+    LOSE("패배"),
+    ;
+
+
+    private final String description;
+
+    GameStatus(String description) {
+        this.description = description;
+    }
+}
