+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/board/cell/Cell.java
@@ -1,0 +1,19 @@
+package cleancode.minesweeper.tobe.minesweeper.board.cell;
+
+public interface Cell {
+
+    boolean isLandMine();
+
+    boolean hasLandMineCount();
+
+    CellSnapshot getSnapshot();
+
+    void flag();
+
+    void open();
+
+    boolean isChecked();
+
+    boolean isOpened();
+
+}
