+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/board/cell/CellSnapshot.java
@@ -1,0 +1,63 @@
+package cleancode.minesweeper.tobe.minesweeper.board.cell;
+
+import java.util.Objects;
+
+public class CellSnapshot {
+
+    private final CellSnapshotStatus status;
+    private final int nearbyLandMineCount;
+
+    private CellSnapshot(CellSnapshotStatus status, int nearbyLandMineCount) {
+        this.status = status;
+        this.nearbyLandMineCount = nearbyLandMineCount;
+    }
+
+    public static CellSnapshot of(CellSnapshotStatus status, int nearbyLandMineCount) {
+        return new CellSnapshot(status, nearbyLandMineCount);
+    }
+
+    public static CellSnapshot ofEmpty() {
+        return of(CellSnapshotStatus.EMPTY, 0);
+    }
+
+    public static CellSnapshot ofFlag() {
+        return of(CellSnapshotStatus.FLAG, 0);
+    }
+
+    public static CellSnapshot ofLandMine() {
+        return of(CellSnapshotStatus.LANE_MINE, 0);
+    }
+
+    public static CellSnapshot ofNumber(int nearbyLandMineCount) {
+        return of(CellSnapshotStatus.NUMBER, nearbyLandMineCount);
+    }
+
+    public static CellSnapshot ofUnchecked() {
+        return of(CellSnapshotStatus.UNCHECKED, 0);
+    }
+
+    public CellSnapshotStatus getStatus() {
+        return status;
+    }
+
+    public int getNearbyLandMineCount() {
+        return nearbyLandMineCount;
+    }
+
+    public boolean isSameStatus(CellSnapshotStatus cellSnapshotStatus) {
+        return this.status == cellSnapshotStatus;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        CellSnapshot snapshot = (CellSnapshot) o;
+        return nearbyLandMineCount == snapshot.nearbyLandMineCount && status == snapshot.status;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(status, nearbyLandMineCount);
+    }
+}
