+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/board/cell/EmptyCell.java
@@ -1,0 +1,48 @@
+package cleancode.minesweeper.tobe.minesweeper.board.cell;
+
+public class EmptyCell implements Cell {
+
+    private final CellState cellState = CellState.initialize();
+
+    @Override
+    public boolean isLandMine() {
+        return false;
+    }
+
+    @Override
+    public boolean hasLandMineCount() {
+        return false;
+    }
+
+    @Override
+    public CellSnapshot getSnapshot() {
+        if (cellState.isOpened()) {
+            return CellSnapshot.ofEmpty();
+        }
+        if (cellState.isFlagged()) {
+            return CellSnapshot.ofFlag();
+        }
+        return CellSnapshot.ofUnchecked();
+    }
+
+    @Override
+    public void flag() {
+        cellState.flag();
+    }
+
+    @Override
+    public void open() {
+        cellState.open();
+    }
+
+    @Override
+    public boolean isChecked() {
+        return cellState.isChecked();
+    }
+
+    @Override
+    public boolean isOpened() {
+        return cellState.isOpened();
+    }
+
+}
