+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/config/GameConfig.java
@@ -1,0 +1,30 @@
+package cleancode.minesweeper.tobe.minesweeper.config;
+
+import cleancode.minesweeper.tobe.minesweeper.gameLevel.GameLevel;
+import cleancode.minesweeper.tobe.minesweeper.io.InputHandler;
+import cleancode.minesweeper.tobe.minesweeper.io.OutputHandler;
+
+public class GameConfig {
+
+    private final GameLevel gameLevel;
+    private final InputHandler inputHandler;
+    private final OutputHandler outputHandler;
+
+    public GameConfig(GameLevel gameLevel, InputHandler inputHandler, OutputHandler outputHandler) {
+        this.gameLevel = gameLevel;
+        this.inputHandler = inputHandler;
+        this.outputHandler = outputHandler;
+    }
+
+    public GameLevel getGameLevel() {
+        return gameLevel;
+    }
+
+    public InputHandler getInputHandler() {
+        return inputHandler;
+    }
+
+    public OutputHandler getOutputHandler() {
+        return outputHandler;
+    }
+}
