+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/exception/GameException.java
@@ -1,0 +1,7 @@
+package cleancode.minesweeper.tobe.minesweeper.exception;
+
+public class GameException extends RuntimeException {
+    public GameException(String message) {
+        super(message);
+    }
+}
