+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/gameLevel/Advanced.java
@@ -1,0 +1,18 @@
+package cleancode.minesweeper.tobe.minesweeper.gameLevel;
+
+public class Advanced implements GameLevel{
+    @Override
+    public int getRowSize() {
+        return 20;
+    }
+
+    @Override
+    public int getColSize() {
+        return 24;
+    }
+
+    @Override
+    public int getLandMineCount() {
+        return 99;
+    }
+}
