+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/gameLevel/GameLevel.java
@@ -1,0 +1,8 @@
+package cleancode.minesweeper.tobe.minesweeper.gameLevel;
+
+public interface GameLevel {
+
+    int getRowSize();
+    int getColSize();
+    int getLandMineCount();
+}
