+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/gameLevel/Middle.java
@@ -1,0 +1,18 @@
+package cleancode.minesweeper.tobe.minesweeper.gameLevel;
+
+public class Middle implements GameLevel{
+    @Override
+    public int getRowSize() {
+        return 14;
+    }
+
+    @Override
+    public int getColSize() {
+        return 18;
+    }
+
+    @Override
+    public int getLandMineCount() {
+        return 40;
+    }
+}
