+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/gameLevel/VeryBeginner.java
@@ -1,0 +1,18 @@
+package cleancode.minesweeper.tobe.minesweeper.gameLevel;
+
+public class VeryBeginner implements GameLevel{
+    @Override
+    public int getRowSize() {
+        return 4;
+    }
+
+    @Override
+    public int getColSize() {
+        return 5;
+    }
+
+    @Override
+    public int getLandMineCount() {
+        return 2;
+    }
+}
