+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/io/InputHandler.java
@@ -1,0 +1,12 @@
+package cleancode.minesweeper.tobe.minesweeper.io;
+
+import cleancode.minesweeper.tobe.minesweeper.board.position.CellPosition;
+import cleancode.minesweeper.tobe.minesweeper.user.UserAction;
+
+public interface InputHandler {
+
+    UserAction getUserActionFromUser();
+
+    CellPosition getCellPositionFromUser();
+
+}
