+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/io/OutputHandler.java
@@ -1,0 +1,24 @@
+package cleancode.minesweeper.tobe.minesweeper.io;
+
+import cleancode.minesweeper.tobe.minesweeper.board.GameBoard;
+import cleancode.minesweeper.tobe.minesweeper.exception.GameException;
+
+public interface OutputHandler {
+
+    void showGameStartComments();
+
+    void showBoard(GameBoard board);
+
+    void showGameWinningComment();
+
+    void showGameLosingComment();
+
+    void showCommentForSelectingCell();
+
+    void showCommentForUserAction();
+
+    void showExceptionMessage(GameException e);
+
+    void showSimpleMessage(String message);
+
+}
