+++ src/main/java/cleancode/minesweeper/tobe/minesweeper/io/sign/CellSignProvidable.java
@@ -1,0 +1,11 @@
+package cleancode.minesweeper.tobe.minesweeper.io.sign;
+
+import cleancode.minesweeper.tobe.minesweeper.board.cell.CellSnapshot;
+
+public interface CellSignProvidable {
+
+    boolean supports(CellSnapshot cellSnapshot);
+
+    String provide(CellSnapshot cellSnapshot);
+
+}
